test/web-platform-tests/to-upstream/html/semantics/tabular-data/the-td-element/class-hierarchy.html

<!DOCTYPE HTML>
<title>HTMLTableCellElement (for &lt;td&gt;)</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/tables.html#the-td-element">
<link rel="author" title="Timothy Gu" href="mailto:timothygu99@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
"use strict";
test(() => {
  const td = document.createElement("td");
  assert_true(td instanceof HTMLTableCellElement, "td instanceof HTMLTableCellElement");
}, "<td> is an instance of HTMLTableCellElement");

test(() => {
  const td = document.createElement("td");
  assert_equals(Object.getPrototypeOf(td), HTMLTableCellElement.prototype,
                "td.__proto__ === HTMLTableCellElement.prototype");
}, "<td> is not an instance of a class inheriting HTMLTableCellElement");
</script>
